Full functionality of cloned clients
Adds full functionality to clone connected clients on multiple clients. Has some issues with syncing; Player A can move and update on Player B's game, but player B is unable to move and update on player A's game after initial creation.

Moved coroutines to use WaitForSecondsRealtime to account for built in TimeScale changes in the game